Remove non-working CI view context menu
