+++ src/main/java/com/jfrog/ide/idea/ui/ComponentsTree.java
@@ -4,15 +4,12 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.JBMenuItem;
 import com.intellij.openapi.ui.JBPopupMenu;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.pom.Navigatable;
 import com.intellij.psi.PsiInvalidElementAccessException;
 import com.intellij.ui.components.JBMenu;
 import com.intellij.ui.treeStructure.Tree;
-import com.jfrog.ide.common.tree.DependencyNode;
 import com.jfrog.ide.idea.exclusion.Excludable;
 import com.jfrog.ide.idea.exclusion.ExclusionUtils;
 import com.jfrog.ide.idea.log.Logger;
-import com.jfrog.ide.idea.navigation.NavigationService;
 import com.jfrog.ide.idea.navigation.NavigationTarget;
 import com.jfrog.ide.idea.ui.menus.ToolbarPopupMenu;
 import org.jetbrains.annotations.NotNull;
@@ -20,11 +17,7 @@ import org.jfrog.build.extractor.scan.DependencyTree;
 
 import javax.swing.*;
 import javax.swing.tree.TreeModel;
-import javax.swing.tree.TreePath;
 import java.awt.event.ActionEvent;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -36,13 +29,12 @@ import java.util.Set;
  * @author yahavi
  */
 public abstract class ComponentsTree extends Tree {
-    private static final String SHOW_IN_PROJECT_DESCRIPTOR = "Show dependency in project descriptor";
+    protected Project project;
+    protected final List<ToolbarPopupMenu> toolbarPopupMenus = new ArrayList<>();
+    protected final JBPopupMenu popupMenu = new JBPopupMenu();
+
     private static final String EXCLUDE_DEPENDENCY = "Exclude dependency";
 
-    protected final List<ToolbarPopupMenu> toolbarPopupMenus = new ArrayList<>();
-    private final JBPopupMenu popupMenu = new JBPopupMenu();
-
-    protected Project project;
 
     public ComponentsTree(@NotNull Project project) {
         super((TreeModel) null);
@@ -59,60 +51,6 @@ public abstract class ComponentsTree extends Tree {
         this.toolbarPopupMenus.add(filterMenu);
     }
 
-    public void addRightClickListener() {
-        MouseListener mouseListener = new MouseAdapter() {
-            @Override
-            public void mousePressed(MouseEvent e) {
-                handleContextMenu(ComponentsTree.this, e);
-            }
-        };
-        addMouseListener(mouseListener);
-    }
-
-    private void handleContextMenu(ComponentsTree tree, MouseEvent e) {
-        if (!e.isPopupTrigger()) {
-            return;
-        }
-        // Event is right-click.
-        TreePath selectedPath = tree.getPathForLocation(e.getX(), e.getY());
-        if (selectedPath == null) {
-            return;
-        }
-        if (selectedPath.getLastPathComponent() instanceof DependencyNode) {
-            createNodePopupMenu((DependencyNode) selectedPath.getLastPathComponent());
-            popupMenu.show(tree, e.getX(), e.getY());
-        }
-    }
-
-
-    private void createNodePopupMenu(DependencyNode selectedNode) {
-        popupMenu.removeAll();
-        NavigationService navigationService = NavigationService.getInstance(project);
-        Set<NavigationTarget> navigationCandidates = navigationService.getNavigation(selectedNode);
-
-        addNodeNavigation(navigationCandidates);
-    }
-
-    private void addNodeNavigation(Set<NavigationTarget> navigationCandidates) {
-        if (navigationCandidates.size() > 1) {
-            addMultiNavigation(navigationCandidates);
-        } else {
-            addSingleNavigation(navigationCandidates.iterator().next());
-        }
-    }
-
-    private void addSingleNavigation(NavigationTarget navigationTarget) {
-        popupMenu.add(createNavigationMenuItem(navigationTarget, SHOW_IN_PROJECT_DESCRIPTOR + " (" + navigationTarget.getComponentName() + ")"));
-    }
-
-    private void addMultiNavigation(Set<NavigationTarget> navigationCandidates) {
-        JMenu multiMenu = new JBMenu();
-        multiMenu.setText(SHOW_IN_PROJECT_DESCRIPTOR);
-        for (NavigationTarget navigationTarget : navigationCandidates) {
-            multiMenu.add(createNavigationMenuItem(navigationTarget, navigationTarget.getComponentName()));
-        }
-        popupMenu.add(multiMenu);
-    }
 
     private String getRelativizedDescriptorPath(NavigationTarget navigationTarget) {
         String pathResult = "";
@@ -133,20 +71,6 @@ public abstract class ComponentsTree extends Tree {
         return pathResult;
     }
 
-    private JMenuItem createNavigationMenuItem(NavigationTarget navigationTarget, String headLine) {
-        return new JBMenuItem(new AbstractAction(headLine) {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                if (!(navigationTarget.getElement() instanceof Navigatable)) {
-                    return;
-                }
-                Navigatable navigatable = (Navigatable) navigationTarget.getElement();
-                if (navigatable.canNavigate()) {
-                    navigatable.navigate(true);
-                }
-            }
-        });
-    }
 
     private void addNodeExclusion(DependencyTree nodeToExclude, Set<NavigationTarget> parentCandidates, DependencyTree affectedNode) {
         if (parentCandidates.size() > 1) {
