+++ src/main/java/com/jfrog/ide/idea/ui/LocalComponentsTree.java
@@ -2,20 +2,35 @@ package com.jfrog.ide.idea.ui;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.ui.JBMenuItem;
+import com.intellij.pom.Navigatable;
+import com.intellij.ui.components.JBMenu;
 import com.jfrog.ide.common.tree.BaseTreeNode;
+import com.jfrog.ide.common.tree.DependencyNode;
 import com.jfrog.ide.common.tree.FileTreeNode;
+import com.jfrog.ide.idea.navigation.NavigationService;
+import com.jfrog.ide.idea.navigation.NavigationTarget;
 import com.jfrog.ide.idea.ui.menus.ToolbarPopupMenu;
 import org.jetbrains.annotations.NotNull;
 
+import javax.swing.*;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
+import javax.swing.tree.TreePath;
+import java.awt.event.ActionEvent;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 /**
  * @author yahavi
  */
 public class LocalComponentsTree extends ComponentsTree {
+    private static final String SHOW_IN_PROJECT_DESCRIPTOR = "Show dependency in project descriptor";
+
     List<FileTreeNode> fileNodes = new ArrayList<>();
 
     public LocalComponentsTree(@NotNull Project project) {
@@ -62,5 +77,75 @@ public class LocalComponentsTree extends ComponentsTree {
         setModel(new DefaultTreeModel(root));
         validate();
         repaint();
+    }
+
+    public void addRightClickListener() {
+        MouseListener mouseListener = new MouseAdapter() {
+            @Override
+            public void mousePressed(MouseEvent e) {
+                handleContextMenu(LocalComponentsTree.this, e);
+            }
+        };
+        addMouseListener(mouseListener);
+    }
+
+    private void handleContextMenu(ComponentsTree tree, MouseEvent e) {
+        if (!e.isPopupTrigger()) {
+            return;
+        }
+        // Event is right-click.
+        TreePath selectedPath = tree.getPathForLocation(e.getX(), e.getY());
+        if (selectedPath == null) {
+            return;
+        }
+        if (selectedPath.getLastPathComponent() instanceof DependencyNode) {
+            createNodePopupMenu((DependencyNode) selectedPath.getLastPathComponent());
+            popupMenu.show(tree, e.getX(), e.getY());
+        }
+    }
+
+
+    private void createNodePopupMenu(DependencyNode selectedNode) {
+        popupMenu.removeAll();
+        NavigationService navigationService = NavigationService.getInstance(project);
+        Set<NavigationTarget> navigationCandidates = navigationService.getNavigation(selectedNode);
+
+        addNodeNavigation(navigationCandidates);
+    }
+
+    private void addNodeNavigation(Set<NavigationTarget> navigationCandidates) {
+        if (navigationCandidates.size() > 1) {
+            addMultiNavigation(navigationCandidates);
+        } else {
+            addSingleNavigation(navigationCandidates.iterator().next());
+        }
+    }
+
+    private void addSingleNavigation(NavigationTarget navigationTarget) {
+        popupMenu.add(createNavigationMenuItem(navigationTarget, SHOW_IN_PROJECT_DESCRIPTOR + " (" + navigationTarget.getComponentName() + ")"));
+    }
+
+    private void addMultiNavigation(Set<NavigationTarget> navigationCandidates) {
+        JMenu multiMenu = new JBMenu();
+        multiMenu.setText(SHOW_IN_PROJECT_DESCRIPTOR);
+        for (NavigationTarget navigationTarget : navigationCandidates) {
+            multiMenu.add(createNavigationMenuItem(navigationTarget, navigationTarget.getComponentName()));
+        }
+        popupMenu.add(multiMenu);
+    }
+
+    private JMenuItem createNavigationMenuItem(NavigationTarget navigationTarget, String headLine) {
+        return new JBMenuItem(new AbstractAction(headLine) {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                if (!(navigationTarget.getElement() instanceof Navigatable)) {
+                    return;
+                }
+                Navigatable navigatable = (Navigatable) navigationTarget.getElement();
+                if (navigatable.canNavigate()) {
+                    navigatable.navigate(true);
+                }
+            }
+        });
     }
 }
